extract listen/set/update handlers
